Update composer packages and check for existence of git dir before cleaning repo in build script.

+++ build.sh
@@ -7,7 +7,10 @@
 ##
 
 # Clean up the working directory (useful when building from local dev files)
-git clean -xdf
+if [ -d ".git" ]
+then
+	git clean -xdf
+fi
 
 # Add composer auth file
 if [ ! -z $COMPOSER_USER ] && [ ! -z $COMPOSER_PASS ]
